pyproject.toml
[tool.poetry]
name = "webappbot"
version = "0.1.0"
description = ""
authors = ["AnatoliyKyznichenko <antonkorodan@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.2"
aiogram = "^3.13.1"
sqlalchemy = "^2.0.36"
asyncpg = "^0.29.0"
python-dotenv = "^1.0.1"
alembic = "^1.13.3"
pydantic-settings = "^2.6.0"
uvicorn = "^0.32.0"
jinja2 = "^3.1.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
